Refactor to use LinkNodesProfiler
